+++ sdk/log/exporter.go
@@ -12,6 +12,12 @@ import (
 )
 
 // Exporter handles the delivery of log records to external receivers.
+//
+// Each of the SDK's built-in processors calls Exporter methods serially and
+// never concurrently. Calls made by different [Processor] instances are not
+// synchronized with each other. Exporter implementations do not need to be
+// safe for concurrent use. An Exporter may be shared between Processors only
+// if it supports concurrent use.
 type Exporter interface {
 	// Export transmits log records to a receiver.
 	//
@@ -27,9 +33,6 @@ type Exporter interface {
 	//
 	// Before modifying a Record, the implementation must use Record.Clone
 	// to create a copy that shares no state with the original.
-	//
-	// Export should never be called concurrently with other Export calls.
-	// However, it may be called concurrently with other methods.
 	Export(ctx context.Context, records []Record) error
 
 	// Shutdown is called when the SDK shuts down. Any cleanup or release of
@@ -40,8 +43,6 @@ type Exporter interface {
 	//
 	// After Shutdown is called, calls to Export, Shutdown, or ForceFlush
 	// should perform no operation and return nil error.
-	//
-	// Shutdown may be called concurrently with itself or with other methods.
 	Shutdown(ctx context.Context) error
 
 	// ForceFlush exports log records to the configured Exporter that have not yet
@@ -49,8 +50,6 @@ type Exporter interface {
 	//
 	// The deadline or cancellation of the passed context must be honored. An
 	// appropriate error should be returned in these situations.
-	//
-	// ForceFlush may be called concurrently with itself or with other methods.
 	ForceFlush(ctx context.Context) error
 }